modify add command feature
rejected new contacts with same name as contacts in current list
and allowed new contacts with same number or email.
phone number required 3 or more numbers.
saves email as inputted

changed as it is common for people to have the same names,
and it is impossible for different people to have the same number or email
numbers in singapore are generally 8 digits
emails are generally in lowercase

modified to accept new contacts with same names as contacts in current
list, reject new contacts with same number or email.
made numbers compulsory to have exactly 8 digits
saves email to lowercase

other info: test files have been affected.
comments have been added to the files for future reference.