modify add command feature
accepted  new contacts with same number or email.
phone number required 3 or more numbers.

it is impossible for different people to have the same number or email
numbers is singapore are generally 8 digits

modified to reject new contacts with same number or email.
made numbers compulsory to have exactly 8 digits.